+++ src/main/java/event/ScoreEvent.java
@@ -20,7 +20,7 @@ package event;
 import javafx.event.Event;
 import javafx.event.EventTarget;
 import javafx.event.EventType;
-import logic.BoardScorer;
+import logic.board.BoardScorer;
 import util.Coords;
 import util.StoneColour;
