+++ src/main/java/logic/board/Board.java
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package logic;
+package logic.board;
 
 import util.Coords;
 import util.StoneColour;
