+++ src/main/java/logic/gametree/GameTree.java
@@ -17,7 +17,7 @@
 
 package logic.gametree;
 
-import logic.Board;
+import logic.board.Board;
 import util.Move;
 
 import java.util.List;
