+++ src/main/java/ui/controller/ScorePaneController.java
@@ -29,7 +29,7 @@ import javafx.scene.control.Label;
 import javafx.scene.control.Separator;
 import javafx.scene.layout.Pane;
 import javafx.scene.layout.VBox;
-import logic.BoardScorer;
+import logic.board.BoardScorer;
 import util.StoneColour;
 
 import java.io.IOException;
