+++ src/main/java/ui/drawer/BoardScoreDrawer.java
@@ -22,8 +22,8 @@ import event.EventBus;
 import event.ScoreEvent;
 import javafx.scene.canvas.Canvas;
 import javafx.scene.canvas.GraphicsContext;
-import logic.BoardScorer;
 import logic.GameHandler;
+import logic.board.BoardScorer;
 import util.CoordProjector;
 import util.Coords;
 import util.DrawCoords;
