+++ src/test/java/logic/board/ChainSetTest.java
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package logic;
+package logic.board;
 
 import org.junit.Test;
